changed: put stream details videodatabase methods in separate file
